migrating machine functions
