+++ lib/charms/mongodb_libs/v0/mongodb_provider.py
@@ -15,8 +15,13 @@ import re
 from collections import namedtuple
 from typing import Optional, Set
 
-import ops.model
-from charms.mongodb_libs.v0.helpers import KEY_FILE, generate_password
+from charms.mongodb_libs.v0.machine_helpers import (
+    auth_enabled,
+    stop_mongod_service,
+    start_mongod_service,
+    update_mongod_service,
+)
+from charms.mongodb_libs.v0.helpers import generate_password
 from charms.mongodb_libs.v0.mongodb import MongoDBConfiguration, MongoDBConnection
 from charms.operator_libs_linux.v1 import systemd
 from ops.charm import RelationBrokenEvent, RelationChangedEvent
@@ -25,7 +30,6 @@ from ops.model import ActiveStatus, BlockedStatus, MaintenanceStatus, Relation
 from pymongo.errors import PyMongoError
 
 from charms.operator_libs_linux.v1 import systemd
-from charms.mongodb_libs.v0.helpers import KEY_FILE
 
 
 # The unique Charmhub library identifier, never change it
@@ -47,8 +51,6 @@ LEGACY_REL_NAME = "obsolete"
 MONGODB_PORT = 27017
 MONGODB_VERSION = "5.0"
 PEER = "database-peers"
-MONGO_USER = "mongodb"
-MONGO_DATA_DIR = "/data/db"
 
 Diff = namedtuple("Diff", "added changed deleted")
 Diff.__doc__ = """
@@ -247,106 +249,6 @@ class MongoDBProvider(Object):
         relations = self.model.relations[LEGACY_REL_NAME]
         return {self._get_username_from_relation_id(relation.id) for relation in relations}
 
-    # # TODO move to machine charm helpers
-    # def auth_enabled(self):
-    #     """Checks if mongod service is running with auth enabled."""
-    #     if not os.path.exists("/lib/systemd/system/mongod.service"):
-    #         return False
-
-    #     with open("/lib/systemd/system/mongod.service", "r") as mongodb_service_file:
-    #         mongodb_service = mongodb_service_file.readlines()
-
-    #     for _, line in enumerate(mongodb_service):
-    #         # ExecStart contains the line with the arguments to start mongod service.
-    #         if "ExecStart" in line and "--auth" in line:
-    #             return True
-
-    #     return False
-
-    # # TODO move to machine charm helpers
-    # def stop_mongod_service(self):
-    #     if not systemd.service_running("mongod.service"):
-    #         return
-
-    #     self.charm.unit.status = MaintenanceStatus("stopping MongoDB")
-    #     logger.debug("stopping mongod.service")
-    #     try:
-    #         systemd.service_stop("mongod.service")
-    #     except systemd.SystemdError as e:
-    #         logger.error("failed to stop mongod.service, error:", e)
-    #         raise
-
-    # # TODO move to machine charm helpers
-    # def update_mongod_service(self, auth: bool):
-    #     mongod_start_args = generate_service_args(auth)
-
-    #     with open("/lib/systemd/system/mongod.service", "r") as mongodb_service_file:
-    #         mongodb_service = mongodb_service_file.readlines()
-
-    #     # replace start command with our parameterized one
-    #     for index, line in enumerate(mongodb_service):
-    #         if "ExecStart" in line:
-    #             mongodb_service[index] = mongod_start_args
-
-    #     # systemd gives files in /etc/systemd/system/ precedence over those in /lib/systemd/system/
-    #     # hence our changed file in /etc will be read while maintaining the original one in /lib.
-    #     with open("/etc/systemd/system/mongod.service", "w") as service_file:
-    #         service_file.writelines(mongodb_service)
-
-    #     # mongod requires permissions to /data/db
-    #     mongodb_user = pwd.getpwnam(MONGO_USER)
-    #     os.chown(MONGO_DATA_DIR, mongodb_user.pw_uid, mongodb_user.pw_gid)
-
-    #     # changes to service files are only applied after reloading
-    #     systemd.daemon_reload()
-
-    # # TODO move to machine charm helpers
-    # def generate_service_args(self, auth: bool) -> str:
-    #     # Construct the mongod startup commandline args for systemd, note that commandline
-    #     # arguments take priority over any user set config file options. User options will be
-    #     # configured in the config file. MongoDB handles this merge of these two options.
-    #     machine_ip = self.charm._unit_ip(self.charm.unit)
-    #     mongod_start_args = [
-    #         "ExecStart=/usr/bin/mongod",
-    #         # bind to localhost and external interfaces
-    #         "--bind_ip",
-    #         f"localhost,{machine_ip}",
-    #         # part of replicaset
-    #         "--replSet",
-    #         f"{self.charm.app.name}",
-    #     ]
-
-    #     if auth:
-    #         mongod_start_args.extend(
-    #             [
-    #                 "--auth",
-    #                 # keyFile used for authenti cation replica set peers, cluster auth, implies user authentication hence we cannot have cluster authentication without user authentication. see: https://www.mongodb.com/docs/manual/reference/configuration-options/#mongodb-setting-security.keyFile
-    #                 # TODO: replace with x509
-    #                 "--clusterAuthMode=keyFile",
-    #                 f"--keyFile={KEY_FILE}",
-    #             ]
-    #         )
-
-    #     mongod_start_args.append("\n")
-    #     mongod_start_args = " ".join(mongod_start_args)
-
-    #     return mongod_start_args
-
-    # # TODO move to machine charm helpers
-    # def start_mongod_service(self):
-    #     if systemd.service_running("mongod.service"):
-    #         return
-
-    #     self.charm.unit.status = MaintenanceStatus("starting MongoDB")
-    #     logger.debug("starting mongod.service")
-    #     try:
-    #         systemd.service_start("mongod.service")
-    #     except systemd.SystemdError as e:
-    #         logger.error("failed to enable mongod.service, error:", e)
-    #         raise
-
-    #     self.charm.unit.status = ActiveStatus()
-
     ################################################################################
     # HELPERS
     ################################################################################
@@ -474,106 +376,8 @@ class MongoDBProvider(Object):
 # TODO move this somewhere appropriate
 
 # # TODO replace functions to not start with "_"
-# update uses of self.charm
-def auth_enabled():
-    """Checks if mongod service is running with auth enabled."""
-    if not os.path.exists("/lib/systemd/system/mongod.service"):
-        return False
 
-    with open("/etc/systemd/system/mongod.service", "r") as mongodb_service_file:
-        mongodb_service = mongodb_service_file.readlines()
-
-    for _, line in enumerate(mongodb_service):
-        # ExecStart contains the line with the arguments to start mongod service.
-        if "ExecStart" in line and "--auth" in line:
-            return True
-
-    return False
-
-
-def stop_mongod_service():
-    if not systemd.service_running("mongod.service"):
-        return
-
-    logger.debug("stopping mongod.service")
-    try:
-        systemd.service_stop("mongod.service")
-    except systemd.SystemdError as e:
-        logger.error("failed to stop mongod.service, error:", e)
-        raise
-
-
-def start_mongod_service():
-    if systemd.service_running("mongod.service"):
-        return
-
-    logger.debug("starting mongod.service")
-    try:
-        systemd.service_start("mongod.service")
-    except systemd.SystemdError as e:
-        logger.error("failed to enable mongod.service, error:", e)
-        raise
-
-
-def update_mongod_service(auth: bool, machine_ip: str, replset: str):
-    mongod_start_args = generate_service_args(auth, machine_ip, replset)
-
-    with open("/lib/systemd/system/mongod.service", "r") as mongodb_service_file:
-        mongodb_service = mongodb_service_file.readlines()
-
-    # replace start command with our parameterized one
-    for index, line in enumerate(mongodb_service):
-        if "ExecStart" in line:
-            mongodb_service[index] = mongod_start_args
-
-    # systemd gives files in /etc/systemd/system/ precedence over those in /lib/systemd/system/
-    # hence our changed file in /etc will be read while maintaining the original one in /lib.
-    with open("/etc/systemd/system/mongod.service", "w") as service_file:
-        service_file.writelines(mongodb_service)
-
-    # mongod requires permissions to /data/db
-    mongodb_user = pwd.getpwnam(MONGO_USER)
-    os.chown(MONGO_DATA_DIR, mongodb_user.pw_uid, mongodb_user.pw_gid)
-
-    # changes to service files are only applied after reloading
-    systemd.daemon_reload()
-
-
-# TODO replace calls of this with machine_ip=self._unit_ip(self.charm.unit)
-def generate_service_args(auth: bool, machine_ip: str, replset: str) -> str:
-    # Construct the mongod startup commandline args for systemd, note that commandline
-    # arguments take priority over any user set config file options. User options will be
-    # configured in the config file. MongoDB handles this merge of these two options.
-    mongod_start_args = [
-        "ExecStart=/usr/bin/mongod",
-        # bind to localhost and external interfaces
-        "--bind_ip",
-        f"localhost,{machine_ip}",
-        # part of replicaset
-        "--replSet",
-        f"{replset}",
-    ]
-
-    if auth:
-        mongod_start_args.extend(
-            [
-                "--auth",
-                # keyFile used for authenti cation replica set peers, cluster auth, implies user authentication hence we cannot have cluster authentication without user authentication. see: https://www.mongodb.com/docs/manual/reference/configuration-options/#mongodb-setting-security.keyFile
-                # TODO: replace with x509
-                "--clusterAuthMode=keyFile",
-                f"--keyFile={KEY_FILE}",
-            ]
-        )
-
-    mongod_start_args.append("\n")
-    mongod_start_args = " ".join(mongod_start_args)
-
-    return mongod_start_args
-
-
-class ApplicationHostNotFoundError(Exception):
-    """Raised when a queried host is not in the application peers or the current host."""
-
-
-class ApplicationHostNotFoundError(Exception):
-    """Raised when a queried host is not in the application peers or the current host."""
+auth_enabled,
+stop_mongod_service,
+start_mongod_service,
+update_mongod_service,
